build.hooks: Use uv for cross builds & installs
When cross compiling create a virtual environment for the build.

Because Nixpkgs builds cross compiled Python in a separate
prefix from the native Python, where the native Python doesn't contain
the sysconfigdata files for the cross Python.
This trips up UV's interpreter discovery scripts which is invoked in isolated mode (-I).

Make the build-host aware of the build-target by aggregating them into a venv.