build: Remove build host Python from cross compiled environment
When installing a Python program with a shebang the shebang points to the native Python.
Because the virtualenv builder symlinks packages to their original path, it means that we retain references to the native Python, bringing it into the runtime closure of a cross compiled application.

By rewriting the shebang to the _target host_ Python at install time we can avoid references to the _build host_ Python in the runtime closure.